// scripts/linkedin/fetch-jobs/data/linkedin-search-enums.ts
export enum ExperienceLevel {
  Intern = "1",
  Entry = "2",
  Mid = "3",
  Senior = "4",
  Director = "5",
  Executive = "6",
}

export enum WorkplaceType {
  OnSite = "1",
  Remote = "2",
  Hybrid = "3",
}

export enum TimePostedRange {
  Last24Hours = "r86400",
  LastWeek = "r604800",
  LastMonth = "r2592000",
  Last3Months = "r7776000",
  Last6Months = "r15552000",
  LastYear = "r31536000",
}

export enum SortBy {
  Relevance = "R",
  MostRecent = "DD",
}